Main menu (#30)
* Add ignoring meta files

Why the hell was this not there already. The other ones wouldn't work on a Unity project. >:(

* Player movement that's limited to screen bounds.

Took a while because the scene is facing performance issues. May be related to 2019.3 if so, may have to revert to earlier version and hope Linux works fine.

* Addressed lint issues

Extracted bool checks to explain values

* Added TextMeshPro

* Main menu

* Remove merge anomoly scene