Abstract out SpaceEntity components
Doing this to build a reliable pattern of objects. 
SpaceEntity forms base of all moving or interactible objects. I'm not entirely sold on this design for all entities in the game and a pure component approach might still be better and cleaner but I'm concerned that splits in functionality between player objects and non-player objects will be duplicating a lot of code.