Enemy ship base (#35)
* Abstract out SpaceEntity components

Doing this to build a reliable pattern of objects. 
SpaceEntity forms base of all moving or interactible objects. I'm not entirely sold on this design for all entities in the game and a pure component approach might still be better and cleaner but I'm concerned that splits in functionality between player objects and non-player objects will be duplicating a lot of code.

* Take hit damage on hit

* Update Unity version

* Make initial enemy spawner

* Give cube health starting HP

* Enemy moves towards target and then stops

* Codacy adjustments

* Clamp player position in own class